api/requirements.txt
requests==2.31.0
python-telegram-bot==21.0.1
telethon==1.36.0
PyJWT==2.10.1
selenium==4.38.0
webdriver-manager==4.0.2
python-dotenv==1.2.1

